+++ lib/hooks/useSession.ts
@@ -9,21 +9,25 @@ const useSession = () => {
   const [hasReports, setHasReports] = useState<boolean | undefined>(undefined);
 
   async function loadSession() {
-    try {
-      const resp = await fetch(`${process.env.NEXT_PUBLIC_API_URL}/auth/session`, {
-        method: "GET",
-        headers: {
-          Authorization: `Bearer ${sessionToken}`
-        }
-      });
+    const resp = await fetch(`${process.env.NEXT_PUBLIC_API_URL}/auth/session`, {
+      method: "GET",
+      headers: {
+        Authorization: `Bearer ${sessionToken}`
+      }
+    });
 
+    if (resp.ok) {
       const data = await resp.json();
 
-      store.setSession({ onboarded: data.is_onboarded, waitlisted: data.is_waitlisted });
+      store.setSession({
+        onboarded: data.is_onboarded,
+        waitlisted: data.is_waitlisted,
+        insightRepoLimit: data.insights_role >= 50 ? 50 : 10
+      });
 
       setHasReports(data.insights_role >= 50);
-    } catch (e) {
-      // show an alert
+    } else {
+    // show an alert
     }
   }